refactor: :truck: sposta file secondo atomi design
